+++ __tests__/server.test.js
@@ -198,6 +198,12 @@ describe("generator-engage:server", () => {
       });
     });
 
+    describe("e2e", () => {
+      it("creates an e2e env file", () => {
+        assert.file("server/src/boot/environments/e2e.js");
+      });
+    });
+
     describe("dotenv", () => {
       it("adds dotenv as a dev dependency", () => {
         expect(json.devDependencies.dotenv).toBeDefined();
@@ -338,15 +344,15 @@ describe("generator-engage:server", () => {
       assert.file("server/src/models/index.js");
     });
 
-    it("creates a boot/model.js", () => {
-      assert.file("server/src/boot/model.js");
+    it("creates a boot/model.cjs", () => {
+      assert.file("server/src/boot/model.cjs");
     });
     it("creates a server/src/db/migration directory", () => {
       assert.file("server/src/db/migrations/migration.stub.cjs");
     });
 
-    it("creates a server/test/utils/truncateModel.js", () => {
-      assert.file("server/test/utils/truncateModel.js");
+    it("creates a server/test/utils/truncateModel.cjs", () => {
+      assert.file("server/test/utils/truncateModel.cjs");
     });
 
     it("adds a migrate:latest script", () => {
@@ -358,6 +364,14 @@ describe("generator-engage:server", () => {
 
     it("adds a migrate:make script", () => {
       expect(json.scripts["migrate:make"]).toBeDefined();
+    });
+
+    it("adds a db:test:migrate script", () => {
+      expect(json.scripts["db:test:migrate"]).toBeDefined();
+    });
+
+    it("adds a db:e2e:migrate script", () => {
+      expect(json.scripts["db:e2e:migrate"]).toBeDefined();
     });
 
     it("adds a console script", () => {