+++ generators/e2e/initE2eOptions.js
@@ -1,0 +1,3 @@
+const initE2eOptions = () => {};
+
+module.exports = initE2eOptions;