+++ generators/e2e/templates/cypress/integration/hello.js
@@ -1,0 +1,10 @@
+/// <reference types="Cypress" />
+
+import { cyan } from "chalk";
+
+describe("hello", () => {
+  it("says hello", () => {
+    cy.visit("/client");
+    cy.contains("Hello");
+  });
+});