+++ generators/server/templates/src/boot.js
@@ -1,7 +1,10 @@
 import "./boot/environments/development.js";
 import "./boot/environments/test.js";
+<% if(options["dbClient"] === "cypress") { %>
+import "./boot/environments/e2e.js"
+<% } %>
 
 <% if(options["dbClient"] === "objection") { %>
-import connection from "./boot/model.js";
+import connection from "./boot/model.cjs";
 export { connection }
 <% } %>