+++ generators/server/templates/src/boot/environments/e2e.js
@@ -1,0 +1,7 @@
+import getNodeEnv from "../../config/getNodeEnv.js";
+
+if (getNodeEnv() === "e2e") {
+  // development specific middlewares here
+  const { default: dotenv } = await import("dotenv");
+  await dotenv.config();
+}