+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - support running webpack bidirectionally, handle for an allow list of client-side routes [#68](https://github.com/LaunchAcademy/generator-engage/issues/68)
 - specify node engine directives in all package.jsons on the basis of the generator runtime
 - generate vscode workspace settings file for color differentiation [#89](https://github.com/LaunchAcademy/generator-engage/issues/89)
+- generate e2e folder with cypress [#28](https://github.com/LaunchAcademy/generator-engage/issues/28)
 
 ### Changed